introduce-output-processor-move: Hand batch-output handling over to OutputProcessor
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod
_handle_batch_output from TokenizerManager; paste into OutputProcessor
(drop @staticmethod, replace ``self: "OutputProcessor"`` -> plain
``self``). Privacy flip rename: _handle_batch_output -> handle_batch_output
(method is now public API of new class). Caller prefix replacement:
``TokenizerManager._handle_batch_output(self.output_processor, ...)`` ->
``self.output_processor.handle_batch_output(...)``.

Refactor chain ID: introduce-output-processor-move
